Complete rewrite of AutoFill to bring it up-to-date with the rest of DataTables for features, style and compatibility.
The same basic features are still present - it performs an Excel like
auto fill using a click and drag handle in the table cells, but how it
actually operates is entirely different. The key change is that how data
is filled is determined by plug-ins - the end user can select what fill
type they want as each plug-in will determine if it can be used for a
fill. This greatly increases the flexibiltiy of AutoFill as it can now
be easily customised.

Also full integration with Editor.